Fix: restructure TestEnvironments into feature groups
